requirements.txt
numpy
ipykernel
pillow
numba
matplotlib